Compatibility changes for numpy>=2 and latest python

+++ udkm1Dsim/simulations/xrays.py
@@ -1361,7 +1361,7 @@ class XrayDyn(Xray):
             rho = self.get_atom_reflection_factor(atom, area, deb_wal_fac)
             tau = self.get_atom_transmission_factor(atom, area, deb_wal_fac)
             # calculate the reflection-transmission matrix
-            H = np.zeros([np.shape(self._qz)[0], np.shape(self._qz)[1], 2, 2], dtype=np.cfloat)
+            H = np.zeros([np.shape(self._qz)[0], np.shape(self._qz)[1], 2, 2], dtype=np.complex128)
             H[:, :, 0, 0] = (1/tau)*(tau**2-rho**2)
             H[:, :, 0, 1] = (1/tau)*(rho)
             H[:, :, 1, 0] = (1/tau)*(-rho)
@@ -1461,7 +1461,7 @@ class XrayDyn(Xray):
 
         """
         phi = self.get_atom_phase_factor(distance)
-        L = np.zeros([np.shape(self._qz)[0], np.shape(self._qz)[1], 2, 2], dtype=np.cfloat)
+        L = np.zeros([np.shape(self._qz)[0], np.shape(self._qz)[1], 2, 2], dtype=np.complex128)
         L[:, :, 0, 0] = np.exp(1j*phi)
         L[:, :, 1, 1] = np.exp(-1j*phi)
         return L
@@ -1638,16 +1638,16 @@ class XrayDynMag(Xray):
 
         self.pol_in_state = pol_in_state
         if (self.pol_in_state == 1):  # circ +
-            self.pol_in = np.array([-np.sqrt(.5), -1j*np.sqrt(.5)], dtype=np.cfloat)
+            self.pol_in = np.array([-np.sqrt(.5), -1j*np.sqrt(.5)], dtype=np.complex128)
         elif (self.pol_in_state == 2):  # circ -
-            self.pol_in = np.array([np.sqrt(.5), -1j*np.sqrt(.5)], dtype=np.cfloat)
+            self.pol_in = np.array([np.sqrt(.5), -1j*np.sqrt(.5)], dtype=np.complex128)
         elif (self.pol_in_state == 3):  # sigma
-            self.pol_in = np.array([1, 0], dtype=np.cfloat)
+            self.pol_in = np.array([1, 0], dtype=np.complex128)
         elif (self.pol_in_state == 4):  # pi
-            self.pol_in = np.array([0, 1], dtype=np.cfloat)
+            self.pol_in = np.array([0, 1], dtype=np.complex128)
         else:  # unpolarized
             self.pol_in_state = 0  # catch any number and set state to 0
-            self.pol_in = np.array([np.sqrt(.5), np.sqrt(.5)], dtype=np.cfloat)
+            self.pol_in = np.array([np.sqrt(.5), np.sqrt(.5)], dtype=np.complex128)
 
         self.disp_message('incoming polarizations set to: {:s}'.format(
             self.polarizations[self.pol_in_state]))
@@ -1665,16 +1665,16 @@ class XrayDynMag(Xray):
 
         self.pol_out_state = pol_out_state
         if (self.pol_out_state == 1):  # circ +
-            self.pol_out = np.array([-np.sqrt(.5), 1j*np.sqrt(.5)], dtype=np.cfloat)
+            self.pol_out = np.array([-np.sqrt(.5), 1j*np.sqrt(.5)], dtype=np.complex128)
         elif (self.pol_out_state == 2):  # circ -
-            self.pol_out = np.array([np.sqrt(.5), 1j*np.sqrt(.5)], dtype=np.cfloat)
+            self.pol_out = np.array([np.sqrt(.5), 1j*np.sqrt(.5)], dtype=np.complex128)
         elif (self.pol_out_state == 3):  # sigma
-            self.pol_out = np.array([1, 0], dtype=np.cfloat)
+            self.pol_out = np.array([1, 0], dtype=np.complex128)
         elif (self.pol_out_state == 4):  # pi
-            self.pol_out = np.array([0, 1], dtype=np.cfloat)
+            self.pol_out = np.array([0, 1], dtype=np.complex128)
         else:  # no analyzer
             self.pol_out_state = 0  # catch any number and set state to 0
-            self.pol_out = np.array([], dtype=np.cfloat)
+            self.pol_out = np.array([], dtype=np.complex128)
 
         self.disp_message('analyzer polarizations set to: {:s}'.format(
             self.polarizations[self.pol_out_state]))
@@ -2479,11 +2479,11 @@ class XrayDynMag(Xray):
              np.sin(mag_gamma),
              np.cos(mag_phi)]
 
-        eps = np.zeros([M, N, 3, 3], dtype=np.cfloat)
-        A = np.zeros([M, N, 4, 4], dtype=np.cfloat)
-        A_phi = np.zeros_like(A, dtype=np.cfloat)
-        P = np.zeros_like(A, dtype=np.cfloat)
-        P_phi = np.zeros_like(A, dtype=np.cfloat)
+        eps = np.zeros([M, N, 3, 3], dtype=np.complex128)
+        A = np.zeros([M, N, 4, 4], dtype=np.complex128)
+        A_phi = np.zeros_like(A, dtype=np.complex128)
+        P = np.zeros_like(A, dtype=np.complex128)
+        P_phi = np.zeros_like(A, dtype=np.complex128)
 
         try:
             molar_density = density/1000/atom.mass_number_a
@@ -2497,11 +2497,11 @@ class XrayDynMag(Xray):
         try:
             cf = atom.get_atomic_form_factor(energy)
         except AttributeError:
-            cf = np.zeros_like(energy, dtype=np.cfloat)
+            cf = np.zeros_like(energy, dtype=np.complex128)
         try:
             mf = atom.get_magnetic_form_factor(energy)
         except AttributeError:
-            mf = np.zeros_like(energy, dtype=np.cfloat)
+            mf = np.zeros_like(energy, dtype=np.complex128)
 
         mag = factor * molar_density * mag_amplitude * mf
         mag = np.tile(mag[:, np.newaxis], [1, N])
@@ -2644,9 +2644,9 @@ class XrayDynMag(Xray):
 
         """
 
-        Ref = np.tile(np.eye(2, 2, dtype=np.cfloat)[np.newaxis, np.newaxis, :, :],
+        Ref = np.tile(np.eye(2, 2, dtype=np.complex128)[np.newaxis, np.newaxis, :, :],
                       (np.size(RT, 0), np.size(RT, 1), 1, 1))
-        Trans = np.tile(np.eye(2, 2, dtype=np.cfloat)[np.newaxis, np.newaxis, :, :],
+        Trans = np.tile(np.eye(2, 2, dtype=np.complex128)[np.newaxis, np.newaxis, :, :],
                         (np.size(RT, 0), np.size(RT, 1), 1, 1))
 
         d = np.divide(1, RT[:, :, 3, 3] * RT[:, :, 2, 2] - RT[:, :, 3, 2] * RT[:, :, 2, 3])
@@ -2672,9 +2672,9 @@ class XrayDynMag(Xray):
         if pol_out.size == 0:
             # no analyzer polarization
             R = np.real(np.matmul(np.square(np.absolute(np.matmul(Ref, pol_in))),
-                        np.array([1, 1], dtype=np.cfloat)))
+                        np.array([1, 1], dtype=np.complex128)))
             T = np.real(np.matmul(np.square(np.absolute(np.matmul(Trans, pol_in))),
-                        np.array([1, 1], dtype=np.cfloat)))
+                        np.array([1, 1], dtype=np.complex128)))
         else:
             R = np.real(np.square(np.absolute(np.matmul(np.matmul(Ref, pol_in), pol_out))))
             T = np.real(np.square(np.absolute(np.matmul(np.matmul(Trans, pol_in), pol_out))))
@@ -2716,7 +2716,7 @@ class XrayDynMag(Xray):
             W (ndarray[float]): roughness matrix.
 
         """
-        W = np.zeros([k_z.shape[0], k_z.shape[1], 4, 4], dtype=np.cfloat)
+        W = np.zeros([k_z.shape[0], k_z.shape[1], 4, 4], dtype=np.complex128)
         rugosp = np.exp(-((k_z + last_k_z)**2) * roughness**2 / 2)
         rugosn = np.exp(-((-k_z + last_k_z)**2) * roughness**2 / 2)
         W[:, :, 0, 0] = rugosn
